+++ lib/yle/aws/role/cli.rb
@@ -20,11 +20,11 @@ module Yle
             o.separator '    account         The account ID or pattern of the role account'
             o.separator '    command         Command to execute with the role. Defaults to launching new shell session.'
             o.separator ''
-            o.integer '-d', '--duration', "Duration for the role credentials. Default: #{Role::DEFAULT_DURATION}"
+            o.integer '-d', '--duration', "Duration for the role credentials. Default: #{Role.default_duration}"
             o.bool '--env', 'Print out environment variables and exit'
             o.bool '-l', '--list', 'Print out all configured account aliases'
             o.bool '-q', '--quiet', 'Be quiet'
-            o.string '-r', '--role', 'Name of the role'
+            o.string '-r', '--role', "Name of the role. Default: '#{Role.default_role_name}'"
             o.separator ''
             o.on '-h', '--help', 'Prints this help' do
               puts o
@@ -39,9 +39,14 @@ module Yle
           @account_name = opts.args.shift
           @command = opts.args
 
-          if !@account_name && !@opts[:list]
-            STDERR.puts @opts
-            exit 64
+          if !@opts[:list]
+            if !@account_name
+              STDERR.puts @opts
+              exit 64
+            elsif !(@opts[:role] || Role.default_role_name)
+              STDERR.puts 'Role name must be passed with `--role` or set in the config'
+              exit 64
+            end
           end
         rescue Slop::Error => e
           STDERR.puts e
@@ -54,12 +59,7 @@ module Yle
             return
           end
 
-          if !role_name
-            STDERR.puts 'Role name must be passed with `--role` or set in the config'
-            exit 64
-          end
-
-          Role.assume_role(account_name, role_name, duration) do |role|
+          Role.assume_role(account_name, opts[:role], opts[:duration]) do |role|
             STDERR.puts("Assumed role #{role.name}") if !opts[:quiet]
 
             if opts[:env]
@@ -89,14 +89,6 @@ module Yle
           ret = system(*cmd)
           STDERR.puts "Failed to execute '#{cmd.first}'" if ret.nil?
           exit(1) if !ret
-        end
-
-        def role_name
-          opts[:role] || Role.config['defaults']['role']
-        end
-
-        def duration
-          opts[:duration] || Role.config['defaults']['duration']
         end
       end
     end