+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.2 / _Not released yet_
 
+- Support default role in the `Role` class
 - Support `require 'yle-aws-role'`
 
 ## 1.0.1 / 2017-05-16